fix: 메인페이지 ui, 핸들러 컴포넌트 분리
